Update Know-Thyself.md
minor proofread throughout; added link to food of the gods; edit to last sentence; and made "T. Mckenna" not in bold- I thought to keep it like the other authors/sources. It also makes the "read thyself" at the bottom stand out. let me know if you want me to change it back.